Fixed results.txt generation
- changed location of test documents and results.txt : now in
(Project).test.documents
- was using new SpamChecker when should of been using original :
original used
- was only classifying documents in ham folder : changed names of files
and mixed into one folder
- was calling SpamChecker.classify(String) using file name instead of
the absolute path : used File.getAbsolutePath() instead
- in the SpamChecker constructor, SpamProbability and HamProbability
were not properly being computed because of integer division : cast to
double was added
- was missing spacing between computed document probabilities : added
- was not adding new line character after each line : added
- was not flushing buffer after each line (not necessary, but prefered)
: added

+++ SpamChecker.java
@@ -34,8 +34,8 @@ public class SpamChecker
 		this.spamDatasetPath = spamDatasetPath;
 		this.hamDocuments    = filterDocuments(hamDatasetPath);
 		this.spamDocuments   = filterDocuments(spamDatasetPath);
-		this.hamProbability  = hamDocuments.size() / (hamDocuments.size() + spamDocuments.size());
-		this.spamProbability = spamDocuments.size() / (hamDocuments.size() + spamDocuments.size());
+		this.hamProbability  = (double)hamDocuments.size() / (hamDocuments.size() + spamDocuments.size());
+		this.spamProbability = (double)spamDocuments.size() / (hamDocuments.size() + spamDocuments.size());
 		this.vocabulary      = this.createVocabulary();
 		computeConditionalProbabilities();
 	}
